chatbot/requirements.txt
langchain>=0.1.17
langchain-openai>=0.0.5
langchain-community>=0.0.19
openai>=1.22.0
python-dotenv>=1.0.1
pydantic>=2.0
tiktoken
